camera calibration ui improvements
